+++ src/components/layout/footer/footer.jsx
@@ -1,6 +1,6 @@
 import styles from './footer.module.css'
-import Social from '../social/social'
-import Logo from '../logo/logo'
+import Social from '../../social/social'
+import Logo from '../../logo/logo'
 
 export default function Footer() {
   return (
